Extract NullsafeQueryRewriter to separate licensed file
- Include MIT license (with permission statement)
- Include Copyright statement by Axel Heer
- Include link to original source
